+++ server/src/helpers/error-handler/index.js
@@ -1,0 +1,16 @@
+const _ = require('lodash');
+
+const errorHandler = err => {
+  const response = {
+    name: err.name,
+    message: err.message,
+  };
+
+  if (err.options) {
+    _.assign(response, { options: err.options });
+  }
+
+  return response;
+};
+
+module.exports = errorHandler;
