+++ server/src/index.js
@@ -2,7 +2,7 @@ require('./configs/database')();
 const express = require('express');
 const morgan = require('morgan');
 const config = require('./config');
-const logger = require('./common/logger');
+const logger = require('./helpers/logger');
 const routes = require('./routes');
 const errorHandler = require('./middlewares/error-handler');
