+++ server/src/middlewares/json-web-token/repository.js
@@ -9,6 +9,7 @@ const jwtConfig = config.jwt;
 const jwtVerify = p.promisify(jwt.verify);
 
 const verifyToken = token => {
+  console.log(token);
   if (!token) {
     return p.reject(Error(errors.MISSING_AUTHORIZATION_HEADER.name));
   }
